backend/pyneumonia/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
brotli==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
django-crum==0.7.9
django-filter==25.2
djangorestframework==3.16.1
fonttools==4.60.1
frozenlist==1.8.0
idna==3.11
inference-sdk==0.59.1
kiwisolver==1.4.9
marshmallow==3.26.1
matplotlib==3.10.7
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.10.0.84
packaging==25.0
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pycparser==2.23
pydyf==0.11.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
supervision==0.26.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
yarl==1.22.0
zopfli==0.4.0
